document_processing/requirements.txt
google-cloud-storage>=2.0.0
google-cloud-vision>=3.1.0
google-cloud-documentai>=2.9.0
google-cloud-bigquery>=3.3.5
vertexai>=0.1.0
pdf2image>=1.16.0
pandas>=1.5.0
openpyxl>=3.0.10
python-dateutil>=2.8.2
functions-framework>=3.0.0